just 2 ruby commands i want to share with others.
the first checks the ruby syntax and shows you the result as a tooltop.

the second executes the current line with ruby and shows you the result,
if there were a error it just shows you the name of the raised exception.

i wanted to publish this earlier but i had problems with ruby-1.6 and
exceptions raised within eval.

i really like to use this excute line thing because you can do small
math convert hex or bin numbers, etc.

the syntax check is bound to ^V, like the css valiador, because they
do the same on different languages, iam not sure if this is ok.




git-svn-id: http://macromates.com/svn/Bundles/trunk/Ruby.tmbundle@340 dfb7d73b-c2ec-0310-8fea-fb051d288c6d