Added support for literals both with and without interpolation, so that constructs such as:   %w(a b c)    and  %W(a b c)   gets colored.
NOTE: I don't think TextMate can balance parentesis, so that  %w(a (b) c) would be possible.

NOTE: In ruby you can make literals with any letter you desire, to support this feature a backref in the end pattern.. refering to the begin pattern... is required.  I am not sure if TextMate supports this.

git-svn-id: http://macromates.com/svn/Bundles/trunk/Ruby.tmbundle@509 dfb7d73b-c2ec-0310-8fea-fb051d288c6d