• Fixed/improved the highlighting of Ruby symbols, I hope it now gets all of them, see http://members.lycos.co.uk/staff/tm/ruby_sym_dbg.rb for with what I tested it, this is a small list of the fixed things:   - ?: and :`` showed up as symbols, I am sure they aren't this.   - :a==, :a!a, :a?a, :a==b, and :a=b claimed to be symbols   - Also :a=> got rendered a bit strange.   - Added all other things which are be symbols, see the comment in line 333 for detailed infos.   - Removed "TODO: deal with these symbols   :"\"x"  :'\'x'" because the quoted symbols should do this ATM. (I tested it)
• Fixed the Rails Syntax because it didn't highlighted Rails keywords (probably because of Simon's commit at r506 in line 89 of the Ruby syntax, I love Blame ;) )


Currently there is only one problem left with the .rhtml files.  Because the default Ruby syntax doesn't include the Rails methods they are not colored.  Also not if I choose Ruby on Rails to be my handler for .rb (it claims to be mode-Ruby, BTW).  It would be nice if TM could remember it if I choose another Syntax to handle mode-Ruby and handle the pattern inclusion in Syntax files this way.  (This is a feature Request :) )

Yes, one could HTML (Ruby) let include Ruby On Rails, but I think this way would be better because so non-railers don't see the methods highlighted in their .rhtmls if they don't want this.


git-svn-id: http://macromates.com/svn/Bundles/trunk/Ruby.tmbundle@530 dfb7d73b-c2ec-0310-8fea-fb051d288c6d