This commit vendors the new rcodetools library in the Ruby bundle.  This is to support the addition of commands based on these tools, like accurate Ruby code completion.
PLEASE REMEMBER THAT THIS LIBRARY HAS ITS OWN LICENSE.  I've carefully reviewed the license and spoke with the developers to ensure that vendoring this code is not a violation.  ANY CHANGES TO THIS CODE MUST ALSO COMPLY WITH THE LICENSE.  (It's better not to modify the code, but instead submit patches back to the developers to avoid this issue entirely.)

Of course, writing commands that shell out to these utilities is fine and, in fact, the whole point of vendoring the library.  I have started us off by rewriting Execute and Update '# =>' Markers in terms of these utilities.


git-svn-id: http://macromates.com/svn/Bundles/trunk/Bundles/Ruby.tmbundle@6425 dfb7d73b-c2ec-0310-8fea-fb051d288c6d