• Update ScriptMate to use new TextMate::Process.run() API.
The previous implementation of scriptmate passed a single string to exec(). This version aims to be safer by passing an array of components that make up the command and as a result not be subject to shell expansion. The Ruby and Python scriptmate implementations either relied upon or catered for shell expansion so they had to be modified.


git-svn-id: http://macromates.com/svn/Bundles/trunk/Bundles/Ruby.tmbundle@9632 dfb7d73b-c2ec-0310-8fea-fb051d288c6d